added fullscreen_x.go
